Migrate to Postgres (#6)

+++ src/entity/user_whitelist.rs
@@ -57,7 +57,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
 
     fn auto_increment() -> bool {
         true
